Removed the code duplication
