+++ frontend/src/components/forms/TextField.tsx
@@ -71,7 +71,9 @@ const TextField: FunctionComponent<Props> = ({
 
   return (
     <div className="w-full">
-      <Label htmlFor={fieldId}>{label}</Label>
+      <Label htmlFor={fieldId} required={otherProps.required}>
+        {label}
+      </Label>
       {error && <InlineError>{error}</InlineError>}
       <TextInput {...otherProps} hasError={!!error} id={fieldId} />
     </div>
